requirements.txt
streamlit
python-dotenv
pandas
numpy
pyarrow
matplotlib
seaborn
plotly
openai